requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pandas>=2.2.0
yfinance>=0.2.54
PyYAML>=6.0.1